Move RevocationInfo to pki/revocation.
