+++ builtin/logical/pki/crl_util.go
@@ -40,13 +40,6 @@ const (
 	unifiedDeltaWALLastRevokedSerial = unifiedDeltaWALPath + deltaWALLastRevokedSerialName
 )
 
-type RevocationInfo struct {
-	CertificateBytes  []byte           `json:"certificate_bytes"`
-	RevocationTime    int64            `json:"revocation_time"`
-	RevocationTimeUTC time.Time        `json:"revocation_time_utc"`
-	CertificateIssuer issuing.IssuerID `json:"issuer_id"`
-}
-
 type revocationRequest struct {
 	RequestedAt time.Time `json:"requested_at"`
 }
@@ -1042,7 +1035,7 @@ func revokeCert(sc *storageContext, config *revocation.CrlConfig, cert *x509.Cer
 	}
 
 	currTime := time.Now()
-	revInfo := RevocationInfo{
+	revInfo := revocation.RevocationInfo{
 		CertificateBytes:  cert.Raw,
 		RevocationTime:    currTime.Unix(),
 		RevocationTimeUTC: currTime.UTC(),
@@ -1870,7 +1863,7 @@ func buildAnyCRLsWithCerts(
 	return warnings, nil
 }
 
-func isRevInfoIssuerValid(revInfo *RevocationInfo, issuerIDCertMap map[issuing.IssuerID]*x509.Certificate) bool {
+func isRevInfoIssuerValid(revInfo *revocation.RevocationInfo, issuerIDCertMap map[issuing.IssuerID]*x509.Certificate) bool {
 	if len(revInfo.CertificateIssuer) > 0 {
 		issuerId := revInfo.CertificateIssuer
 		if _, issuerExists := issuerIDCertMap[issuerId]; issuerExists {
@@ -1881,7 +1874,7 @@ func isRevInfoIssuerValid(revInfo *RevocationInfo, issuerIDCertMap map[issuing.I
 	return false
 }
 
-func associateRevokedCertWithIsssuer(revInfo *RevocationInfo, revokedCert *x509.Certificate, issuerIDCertMap map[issuing.IssuerID]*x509.Certificate) bool {
+func associateRevokedCertWithIsssuer(revInfo *revocation.RevocationInfo, revokedCert *x509.Certificate, issuerIDCertMap map[issuing.IssuerID]*x509.Certificate) bool {
 	for issuerId, issuerCert := range issuerIDCertMap {
 		if bytes.Equal(revokedCert.RawIssuer, issuerCert.RawSubject) {
 			if err := revokedCert.CheckSignatureFrom(issuerCert); err == nil {
@@ -1922,7 +1915,7 @@ func getLocalRevokedCertEntries(sc *storageContext, issuerIDCertMap map[issuing.
 			continue
 		}
 
-		var revInfo RevocationInfo
+		var revInfo revocation.RevocationInfo
 		revokedEntry, err := sc.Storage.Get(sc.Context, revokedPath+serial)
 		if err != nil {
 			return nil, nil, errutil.InternalError{Err: fmt.Sprintf("unable to fetch revoked cert with serial %s: %s", serial, err)}
